third_party/WebKit/LayoutTests/http/tests/devtools/profiler/heap-snapshot-weak-dominator-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that weak references are ignored when dominators are calculated and that weak references won't affect object's retained size.

Profiler was enabled.
Detailed heap profiles were enabled.

Running: testWeakReferencesDoNotAffectRetainedSize

Profiler was disabled.
